Kill clean all remaining processes

+++ util/kill_all_processes.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-for a in $(ps -ax |grep lingi2142 | awk '{ print $1; }')
+for a in $(ps -ax |grep lingi2142 |grep -v 'makefile\|kill_all_processes\|cleanup\|grep' | awk '{ print $1; }')
 do
- sudo kill $a 
-done 
+if ps -p $a > /dev/null; then sudo kill $a; fi
+done
